Move: config 클래스 리패키징
